Slightly refactored cert
