+++ cert/lib/cert.rb
@@ -1,18 +1,4 @@
 require_relative 'cert/runner'
 require_relative 'cert/options'
 
-require 'fastlane_core/helper'
-require 'fastlane_core/ui/ui'
-
-module Cert
-  # Use this to just setup the configuration attribute and set it later somewhere else
-  class << self
-    attr_accessor :config
-  end
-
-  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
-  UI = FastlaneCore::UI
-  ROOT = Pathname.new(File.expand_path('../..', __FILE__))
-
-  ENV['FASTLANE_TEAM_ID'] ||= ENV["CERT_TEAM_ID"]
-end
+require_relative 'cert/module'
